+++ packages/bpk-component-calendar/src/BpkCalendarContainer-test.js
@@ -38,10 +38,11 @@ describe('BpkCalendarContainer', () => {
 
     const grid = calendar.find('BpkCalendarGridTransition');
     const nav = calendar.find('BpkCalendarNav');
+    const eventStub = { persist: jest.fn() };
 
     expect(grid.prop('month')).toEqual(new Date(2010, 1, 1));
 
-    nav.prop('onMonthChange')(new Date(2010, 2, 1));
+    nav.prop('onMonthChange')(eventStub, { month: new Date(2010, 2, 1) });
     expect(grid.prop('month')).toEqual(new Date(2010, 2, 1));
   });
 
@@ -99,6 +100,7 @@ describe('BpkCalendarContainer', () => {
 
   it('should move focus on keyboard input', () => {
     const preventDefault = jest.fn();
+    const persist = jest.fn();
     const origin = new Date(2010, 2, 1);
 
     const calendar = mount(<BpkCalendarContainer
@@ -114,40 +116,68 @@ describe('BpkCalendarContainer', () => {
 
     expect(calendar.state('focusedDate')).toEqual(origin);
 
-    calendar.instance().handleDateKeyDown({ key: 'S', preventDefault });
+    calendar.instance().handleDateKeyDown({ key: 'S', preventDefault, persist });
     expect(calendar.state('focusedDate')).toEqual(origin);
     expect(preventDefault.mock.calls.length).toEqual(0);
 
-    calendar.instance().handleDateKeyDown({ key: 'ArrowRight', preventDefault });
+    calendar.instance().handleDateKeyDown({ key: 'ArrowRight', preventDefault, persist });
     expect(calendar.state('focusedDate')).toEqual(addDays(origin, 1));
     expect(preventDefault.mock.calls.length).toEqual(1);
 
-    calendar.instance().handleDateKeyDown({ key: 'ArrowDown', preventDefault });
+    calendar.instance().handleDateKeyDown({ key: 'ArrowDown', preventDefault, persist });
     expect(calendar.state('focusedDate')).toEqual(addDays(origin, 8));
     expect(preventDefault.mock.calls.length).toEqual(2);
 
-    calendar.instance().handleDateKeyDown({ key: 'ArrowLeft', preventDefault });
+    calendar.instance().handleDateKeyDown({ key: 'ArrowLeft', preventDefault, persist });
     expect(calendar.state('focusedDate')).toEqual(addDays(origin, 7));
     expect(preventDefault.mock.calls.length).toEqual(3);
 
-    calendar.instance().handleDateKeyDown({ key: 'ArrowUp', preventDefault });
+    calendar.instance().handleDateKeyDown({ key: 'ArrowUp', preventDefault, persist });
     expect(calendar.state('focusedDate')).toEqual(origin);
     expect(preventDefault.mock.calls.length).toEqual(4);
 
-    calendar.instance().handleDateKeyDown({ key: 'End', preventDefault });
+    calendar.instance().handleDateKeyDown({ key: 'End', preventDefault, persist });
     expect(calendar.state('focusedDate')).toEqual(addDays(origin, 14));
     expect(preventDefault.mock.calls.length).toEqual(5);
 
-    calendar.instance().handleDateKeyDown({ key: 'Home', preventDefault });
+    calendar.instance().handleDateKeyDown({ key: 'Home', preventDefault, persist });
     expect(calendar.state('focusedDate')).toEqual(origin);
     expect(preventDefault.mock.calls.length).toEqual(6);
 
-    calendar.instance().handleDateKeyDown({ key: 'PageDown', preventDefault });
+    calendar.instance().handleDateKeyDown({ key: 'PageDown', preventDefault, persist });
     expect(calendar.state('focusedDate')).toEqual(addDays(origin, 14));
     expect(preventDefault.mock.calls.length).toEqual(7);
 
-    calendar.instance().handleDateKeyDown({ key: 'PageUp', preventDefault });
+    calendar.instance().handleDateKeyDown({ key: 'PageUp', preventDefault, persist });
     expect(calendar.state('focusedDate')).toEqual(addDays(origin, -14));
     expect(preventDefault.mock.calls.length).toEqual(8);
+  });
+
+  it('should change month on keyboard nav across month boundary', () => {
+    const preventDefault = jest.fn();
+    const persist = jest.fn();
+    const onMonthChange = jest.fn();
+    const origin = new Date(2010, 1, 27);
+
+    const calendar = mount(<BpkCalendarContainer
+      formatMonth={formatMonth}
+      formatDateFull={formatDateFull}
+      daysOfWeek={weekDays}
+      changeMonthLabel="Change month"
+      id="myCalendar"
+      minDate={new Date(2010, 1, 1)}
+      maxDate={new Date(2010, 2, 30)}
+      selectedDate={origin}
+      onMonthChange={onMonthChange}
+    />);
+
+    calendar.instance().handleDateKeyDown({ key: 'ArrowRight', preventDefault, persist });
+    expect(calendar.state('focusedDate')).toEqual(addDays(origin, 1));
+    expect(preventDefault.mock.calls.length).toEqual(1);
+    expect(onMonthChange.mock.calls.length).toEqual(0);
+
+    calendar.instance().handleDateKeyDown({ key: 'ArrowRight', preventDefault, persist });
+    expect(onMonthChange.mock.calls.length).toEqual(1);
+    expect(onMonthChange.mock.calls[0][1]).toEqual({ month: new Date(2010, 2, 1), source: 'GRID' });
   });
 });