+++ packages/bpk-component-calendar/src/BpkCalendarNav-test.js
@@ -34,7 +34,7 @@ describe('BpkCalendarNav', () => {
   });
 
   it('should call the onMonthChange callback when nudging/selecting month', () => {
-    const onMonthChange = jest.fn(); // TODO bug here
+    const onMonthChange = jest.fn();
 
     const nav = shallow(<BpkCalendarNav
       month={new Date(2010, 1, 1)}
@@ -49,18 +49,24 @@ describe('BpkCalendarNav', () => {
     expect(onMonthChange.mock.calls.length).toBe(0);
 
     // Previous month
-    nav.find('button').at(0).simulate('click');
+    const prevEventStub = { persist: jest.fn() };
+    nav.find('button').at(0).simulate('click', prevEventStub);
     expect(onMonthChange.mock.calls.length).toBe(1);
-    expect(onMonthChange.mock.calls[0][0]).toEqual(new Date(2010, 0, 1));
+    expect(onMonthChange.mock.calls[0][0]).toEqual(prevEventStub);
+    expect(onMonthChange.mock.calls[0][1]).toEqual({ month: new Date(2010, 0, 1), source: 'PREV' });
 
     // Next month
-    nav.find('button').at(1).simulate('click');
+    const nextEventStub = { persist: jest.fn() };
+    nav.find('button').at(1).simulate('click', nextEventStub);
     expect(onMonthChange.mock.calls.length).toBe(2);
-    expect(onMonthChange.mock.calls[1][0]).toEqual(new Date(2010, 2, 1));
+    expect(onMonthChange.mock.calls[1][0]).toEqual(nextEventStub);
+    expect(onMonthChange.mock.calls[1][1]).toEqual({ month: new Date(2010, 2, 1), source: 'NEXT' });
 
     // Select month
-    nav.find('BpkSelect').simulate('change', { target: { value: '2010-03-01' } });
+    const selectEventStub = { target: { value: '2010-03-01' }, persist: jest.fn() };
+    nav.find('BpkSelect').simulate('change', selectEventStub);
     expect(onMonthChange.mock.calls.length).toBe(3);
-    expect(onMonthChange.mock.calls[2][0]).toEqual(new Date(2010, 2, 1));
+    expect(onMonthChange.mock.calls[2][0]).toEqual(selectEventStub);
+    expect(onMonthChange.mock.calls[2][1]).toEqual({ month: new Date(2010, 2, 1), source: 'SELECT' });
   });
 });